Move event to Events page and add news feed
